+++ CHANGELOG.md
@@ -6,6 +6,27 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [0.2.0] - 2021-08-07
+### Added
+- CLI commands to use the processor [#1]
+    - process
+        - title
+        - name
+        - destination
+        - file *[to process]*
+        - css theme [#9]
+- Documentation
+    - CLI
+    - Typo feature
+- Themes
+    - RPG [#8]
+        
+### Fixed
+- Management of
+    - code blocks
+    - checkbox list
+- Default CSS style
+
 ## [0.1.0] - 2021-07-31
 ### Added
 - Documentation